Add Lab Access calendar
I'm moving the direct calendar links out of the onboarding card in https://github.com/neuropoly/onboarding/pull/127 to make maintenance easier.

I think this Lab Access calendar was mostly for COVID contact tracing, n'est pas? But it's still active so until we delete it we should have it documented.